# back-end/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Documentación del API del Sistema Flotillas",
    "description": "Definición de servicios del Sistema Flotillas de Grupo Andrade"
  },
  "host": "/",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/catalogo/traslados/removeUbicacionTraslados/{idLugar}": {
      "parameters": [
        {
          "name": "idLugar",
          "in": "path",
          "required": true,
          "description": "Identificador del Almacen o Garage",
          "type": "string"
        }
      ],
      "delete": {
        "tags": [
          "Catálogo - Traslados"
        ],
        "summary": "Eliminar regla aprobacion de margen de utilidad",
        "operationId": "removeAprobacionUtilidadDireccion",
        "responses": {
          "200": {
            "description": "eliminar un Almacen o Garage",
            "schema": {
              "$ref": "/schemas/catalogo/traslados/ubicacion_traslado_array.schema.json"
            }
          }
        }
      }
    },
...
    "/aprobacion/removeAprobacionUtilidadDireccion/{idDireccionFlotillas}/{idUsuario}": {
      "parameters": [
        {
          "name": "idDireccionFlotillas",
          "in": "path",
          "required": true,
          "description": "Identificador de la Direccion",
          "type": "string"
        },
        {
          "name": "idUsuario",
          "in": "path",
          "required": true,
          "description": "Identificador del Usuario",
          "type": "string"
        }
      ],
      "delete": {
        "tags": [
          "Aprobación - Utilidad"
        ],
        "summary": "Eliminar regla aprobacion de margen de utilidad",
        "operationId": "removeAprobacionUtilidadDireccion",
        "responses": {
          "200": {
            "description": "eliminar una reglar de aprobacion",
            "schema": {
              "$ref": "/schemas/aprobacion/aprobacionUtilidadDireccion_array.schema.json"
            }
          }
        }
      }
    },
    "/aprobacion/insertAprobacionUtilidadDireccion": {
      "post": {
        "tags": [
          "Aprobación - Utilidad"
        ],
        "summary": "Inserta un aprobacion de Unidades Direccion",
        "operationId": "insertAprobacionUtilidadDireccion",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Json de la Aprobacion Credito",
            "required": true,
            "schema": {
              "$ref": "/schemas/aprobacion/aprobacionUtilidadDireccion_array.schema.json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Insertar una nueva cotizacion",
            "schema": {
              "$ref": "/schemas/aprobacion/aprobacionUnidadDireccion_array.schema.json"
            }
          }
        }
      }
    },
...
    "/aprobacion/getAprobacionUtilidadByUsuario/{idUsuario}": {
      "parameters": [
        {
          "name": "idDireccionFlotillas",
          "in": "path",
          "required": true,
          "description": "Identificador de la Direccion",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Aprobación - Utilidad"
        ],
        "summary": "Obtiene aprobaciones de credito por direccion",
        "operationId": "getAprobacionUtilidadByUsuario",
        "responses": {
          "200": {
            "description": "Insertar una nueva cotizacion",
            "schema": {
              "$ref": "/schemas/aprobacion/aprobacionUtilidadDireccion_array.schema.json"
            }
          }
        }
      }
    },
    "/aprobacion/getAprobacionUtilidadByDireccion/{idDireccionFlotillas}": {
      "parameters": [
        {
          "name": "idDireccionFlotillas",
          "in": "path",
          "required": true,
          "description": "Identificador de la Direccion",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Aprobación - Utilidad"
        ],
        "summary": "Obtiene aprobaciones de credito por direccion",
        "operationId": "getAprobacionUtilidadByDireccion",
        "responses": {
          "200": {
            "description": "Insertar una nueva cotizacion",
            "schema": {
              "$ref": "/schemas/aprobacion/aprobacionUtilidadDireccion_array.schema.json"
            }
          }
        }
      }
    },
...
  }
}